+++ lib/service/content.dart
@@ -40,7 +40,7 @@ import 'package:uuid/uuid.dart';
 //import 'package:flutter/services.dart' show rootBundle;
 
 // Content service does rely on Service initialization API so it does not override service interfaces and is not registered in Services.
-class Content with Service, NotificationsListener, ContentItemCategoryClient {
+class Content with Service, NotificationsListener implements ContentItemCategoryClient {
 
   static const String notifyContentItemsChanged           = "edu.illinois.rokwire.content.content_items.changed";
   static const String notifyContentAttributesChanged      = "edu.illinois.rokwire.content.attributes.changed";
@@ -262,6 +262,9 @@ class Content with Service, NotificationsListener, ContentItemCategoryClient {
     Map<String, dynamic>? contentItems = await loadContentItems([category]);
     return (contentItems != null) ? contentItems[category] : null;
   }
+
+  Future<void> refreshContentItems() async =>
+    updateContentItemsFromNet();
 
   @protected
   Future<void> updateContentItemsFromNet() async {